+++ jason/src/peer/conn.rs
@@ -1,10 +1,11 @@
 use std::{cell::RefCell, rc::Rc};
 
+use derive_more::Display;
 use medea_client_api_proto::IceServer;
-use thiserror::*;
+use tracerr::Traced;
 use wasm_bindgen_futures::JsFuture;
 use web_sys::{
-    RtcConfiguration, RtcIceCandidateInit,
+    Event, RtcConfiguration, RtcIceCandidateInit, RtcIceConnectionState,
     RtcPeerConnection as SysRtcPeerConnection, RtcPeerConnectionIceEvent,
     RtcRtpTransceiver, RtcRtpTransceiverDirection, RtcRtpTransceiverInit,
     RtcSdpType, RtcSessionDescription, RtcSessionDescriptionInit,
@@ -13,7 +14,7 @@ use web_sys::{
 
 use crate::{
     media::TrackConstraints,
-    utils::{EventListener, WasmErr},
+    utils::{EventListener, EventListenerBindError, JsCaused, JsError},
 };
 
 use super::ice_server::RtcIceServers;
@@ -119,24 +120,51 @@ pub enum SdpType {
 /// [RTCPeerConnection][1] and event handlers setting errors.
 ///
 /// [1]: https://www.w3.org/TR/webrtc/#dom-rtcpeerconnection.
-#[derive(Debug, Error)]
-pub enum Error {
-    #[error("failed to add ICE candidate: {0}")]
-    AddIceCandidate(WasmErr),
-    #[error("failed to create SDP answer: {0}")]
-    CreateAnswer(WasmErr),
-    #[error("failed to create PeerConnection: {0}")]
-    CreatePeer(WasmErr),
-    #[error("failed to create SDP offer: {0}")]
-    CreateOffer(WasmErr),
-    #[error("failed to set handler for RtcPeerConnectionIceEvent: {0}")]
-    SetHandlerIceEvent(WasmErr),
-    #[error("failed to set handler for RtcTrackEvent: {0}")]
-    SetHandlerTrackEvent(WasmErr),
-    #[error("failed to set local SDP description: {0}")]
-    SetLocalDescription(WasmErr),
-    #[error("failed to set remote SDP description: {0}")]
-    SetRemoteDescription(WasmErr),
+#[derive(Debug, Display, JsCaused)]
+pub enum RTCPeerConnectionError {
+    /// Occurs when cannot adds new remote candidate to the
+    /// [RTCPeerConnection][1]'s remote description.
+    ///
+    /// [1]: https://www.w3.org/TR/webrtc/#dom-rtcpeerconnection.
+    #[display(fmt = "Failed to add ICE candidate: {}", _0)]
+    AddIceCandidateFailed(JsError),
+
+    /// Occurs when cannot obtains [SDP answer][`SdpType::Answer`] from
+    /// the underlying [RTCPeerConnection][`SysRtcPeerConnection`].
+    #[display(fmt = "Failed to create SDP answer: {}", _0)]
+    CreateAnswerFailed(JsError),
+
+    /// Occurs when a new [`RtcPeerConnection`] cannot be created.
+    #[display(fmt = "Failed to create PeerConnection: {}", _0)]
+    PeerCreationError(JsError),
+
+    /// Occurs when cannot obtains [SDP offer][`SdpType::Offer`] from
+    /// the underlying [RTCPeerConnection][`SysRtcPeerConnection`]
+    #[display(fmt = "Failed to create SDP offer: {}", _0)]
+    CreateOfferFailed(JsError),
+
+    /// Occurs when handler failed to bind to some [`RTCPeerConnection`] event.
+    /// Not really supposed to ever happen.
+    #[display(fmt = "Failed to bind to RTCPeerConnection event: {}", _0)]
+    PeerConnectionEventBindFailed(EventListenerBindError),
+
+    /// Occurs if the local description associated with the
+    /// [`RTCPeerConnection`] cannot be changed.
+    #[display(fmt = "Failed to set local SDP description: {}", _0)]
+    SetLocalDescriptionFailed(JsError),
+
+    /// Occurs if the description of the remote end of the
+    /// [`RTCPeerConnection`] cannot be changed.
+    #[display(fmt = "Failed to set remote SDP description: {}", _0)]
+    SetRemoteDescriptionFailed(JsError),
+}
+
+type Result<T> = std::result::Result<T, Traced<RTCPeerConnectionError>>;
+
+impl From<EventListenerBindError> for RTCPeerConnectionError {
+    fn from(err: EventListenerBindError) -> Self {
+        Self::PeerConnectionEventBindFailed(err)
+    }
 }
 
 /// Representation of [RTCPeerConnection][1].
@@ -160,6 +188,15 @@ pub struct RtcPeerConnection {
         Option<EventListener<SysRtcPeerConnection, RtcPeerConnectionIceEvent>>,
     >,
 
+    /// [`iceconnectionstatechange`][2] callback of [RTCPeerConnection][1],
+    /// fires whenever [ICE connection state][3] changes.
+    ///
+    /// [1]: https://w3.org/TR/webrtc/#rtcpeerconnection-interface
+    /// [2]: https://w3.org/TR/webrtc/#event-iceconnectionstatechange
+    /// [3]: https://w3.org/TR/webrtc/#dfn-ice-connection-state
+    on_ice_connection_state_changed:
+        RefCell<Option<EventListener<SysRtcPeerConnection, Event>>>,
+
     /// [`ontrack`][2] callback of [RTCPeerConnection][1] to handle
     /// [`track`][3] event. It fires when [RTCPeerConnection][1] receives
     /// new [MediaStreamTrack][4] from remote peer.
@@ -174,7 +211,7 @@ pub struct RtcPeerConnection {
 
 impl RtcPeerConnection {
     /// Instantiates new [`RtcPeerConnection`].
-    pub fn new<I>(ice_servers: I) -> Result<Self, Error>
+    pub fn new<I>(ice_servers: I) -> Result<Self>
     where
         I: IntoIterator<Item = IceServer>,
     {
@@ -183,11 +220,13 @@ impl RtcPeerConnection {
         peer_conf.ice_servers(&RtcIceServers::from(ice_servers));
         let peer = SysRtcPeerConnection::new_with_configuration(&peer_conf)
             .map_err(Into::into)
-            .map_err(Error::CreatePeer)?;
+            .map_err(RTCPeerConnectionError::PeerCreationError)
+            .map_err(tracerr::wrap!())?;
 
         Ok(Self {
             peer: Rc::new(peer),
             on_ice_candidate: RefCell::new(None),
+            on_ice_connection_state_changed: RefCell::new(None),
             on_track: RefCell::new(None),
         })
     }
@@ -197,7 +236,7 @@ impl RtcPeerConnection {
     ///
     /// [1]: https://www.w3.org/TR/webrtc/#rtctrackevent
     /// [2]: https://www.w3.org/TR/webrtc/#dom-rtcpeerconnection-ontrack
-    pub fn on_track<F>(&self, f: Option<F>) -> Result<(), Error>
+    pub fn on_track<F>(&self, f: Option<F>) -> Result<()>
     where
         F: 'static + FnMut(RtcTrackEvent),
     {
@@ -215,7 +254,7 @@ impl RtcPeerConnection {
                             f(msg);
                         },
                     )
-                    .map_err(Error::SetHandlerTrackEvent)?,
+                    .map_err(tracerr::map_from_and_wrap!())?,
                 );
             }
         }
@@ -227,7 +266,7 @@ impl RtcPeerConnection {
     ///
     /// [1]: https://www.w3.org/TR/webrtc/#dom-rtcpeerconnectioniceevent
     /// [2]: https://www.w3.org/TR/webrtc/#dom-rtcpeerconnection-onicecandidate
-    pub fn on_ice_candidate<F>(&self, f: Option<F>) -> Result<(), Error>
+    pub fn on_ice_candidate<F>(&self, f: Option<F>) -> Result<()>
     where
         F: 'static + FnMut(IceCandidate),
     {
@@ -255,7 +294,37 @@ impl RtcPeerConnection {
                             }
                         },
                     )
-                    .map_err(Error::SetHandlerIceEvent)?,
+                    .map_err(tracerr::map_from_and_wrap!())?,
+                );
+            }
+        }
+        Ok(())
+    }
+
+    /// Sets handler for [`iceconnectionstatechange`][1] event.
+    ///
+    /// [1]: https://www.w3.org/TR/webrtc/#event-iceconnectionstatechange
+    pub fn on_ice_connection_state_change<F>(&self, f: Option<F>) -> Result<()>
+    where
+        F: 'static + FnMut(RtcIceConnectionState),
+    {
+        let mut on_ice_connection_state_changed =
+            self.on_ice_connection_state_changed.borrow_mut();
+        match f {
+            None => {
+                on_ice_connection_state_changed.take();
+            }
+            Some(mut f) => {
+                let peer = Rc::clone(&self.peer);
+                on_ice_connection_state_changed.replace(
+                    EventListener::new_mut(
+                        Rc::clone(&self.peer),
+                        "iceconnectionstatechange",
+                        move |_| {
+                            f(peer.ice_connection_state());
+                        },
+                    )
+                    .map_err(tracerr::map_from_and_wrap!())?,
                 );
             }
         }
@@ -272,7 +341,7 @@ impl RtcPeerConnection {
         candidate: &str,
         sdp_m_line_index: Option<u16>,
         sdp_mid: &Option<String>,
-    ) -> Result<(), Error> {
+    ) -> Result<()> {
         let mut cand_init = RtcIceCandidateInit::new(&candidate);
         cand_init
             .sdp_m_line_index(sdp_m_line_index)
@@ -284,7 +353,8 @@ impl RtcPeerConnection {
         )
         .await
         .map_err(Into::into)
-        .map_err(Error::AddIceCandidate)?;
+        .map_err(RTCPeerConnectionError::AddIceCandidateFailed)
+        .map_err(tracerr::wrap!())?;
         Ok(())
     }
 
@@ -293,13 +363,14 @@ impl RtcPeerConnection {
     /// description.
     ///
     /// Should be called whenever remote description has been changed.
-    pub async fn create_and_set_answer(&self) -> Result<String, Error> {
+    pub async fn create_and_set_answer(&self) -> Result<String> {
         let peer: Rc<SysRtcPeerConnection> = Rc::clone(&self.peer);
 
         let answer = JsFuture::from(self.peer.create_answer())
             .await
             .map_err(Into::into)
-            .map_err(Error::CreateAnswer)?;
+            .map_err(RTCPeerConnectionError::CreateAnswerFailed)
+            .map_err(tracerr::wrap!())?;
         let answer = RtcSessionDescription::from(answer).sdp();
 
         let mut desc = RtcSessionDescriptionInit::new(RtcSdpType::Answer);
@@ -308,7 +379,8 @@ impl RtcPeerConnection {
         JsFuture::from(peer.set_local_description(&desc))
             .await
             .map_err(Into::into)
-            .map_err(Error::SetLocalDescription)?;
+            .map_err(RTCPeerConnectionError::SetLocalDescriptionFailed)
+            .map_err(tracerr::wrap!())?;
 
         Ok(answer)
     }
@@ -319,13 +391,14 @@ impl RtcPeerConnection {
     ///
     /// Should be called after local tracks changes, which require
     /// renegotiation.
-    pub async fn create_and_set_offer(&self) -> Result<String, Error> {
+    pub async fn create_and_set_offer(&self) -> Result<String> {
         let peer: Rc<SysRtcPeerConnection> = Rc::clone(&self.peer);
 
         let create_offer = JsFuture::from(peer.create_offer())
             .await
             .map_err(Into::into)
-            .map_err(Error::CreateOffer)?;
+            .map_err(RTCPeerConnectionError::CreateOfferFailed)
+            .map_err(tracerr::wrap!())?;
         let offer = RtcSessionDescription::from(create_offer).sdp();
 
         let mut desc = RtcSessionDescriptionInit::new(RtcSdpType::Offer);
@@ -334,7 +407,8 @@ impl RtcPeerConnection {
         JsFuture::from(peer.set_local_description(&desc))
             .await
             .map_err(Into::into)
-            .map_err(Error::SetLocalDescription)?;
+            .map_err(RTCPeerConnectionError::SetLocalDescriptionFailed)
+            .map_err(tracerr::wrap!())?;
 
         Ok(offer)
     }
@@ -344,10 +418,7 @@ impl RtcPeerConnection {
     /// [offer][`SdpType::Offer`] or [answer][`SdpType::Answer`].
     ///
     /// Changes the local media state.
-    pub async fn set_remote_description(
-        &self,
-        sdp: SdpType,
-    ) -> Result<(), Error> {
+    pub async fn set_remote_description(&self, sdp: SdpType) -> Result<()> {
         let description = match sdp {
             SdpType::Offer(offer) => {
                 let mut desc =
@@ -366,7 +437,8 @@ impl RtcPeerConnection {
         JsFuture::from(self.peer.set_remote_description(&description))
             .await
             .map_err(Into::into)
-            .map_err(Error::SetRemoteDescription)?;
+            .map_err(RTCPeerConnectionError::SetRemoteDescriptionFailed)
+            .map_err(tracerr::wrap!())?;
 
         Ok(())
     }
@@ -425,6 +497,7 @@ impl Drop for RtcPeerConnection {
     fn drop(&mut self) {
         self.on_track.borrow_mut().take();
         self.on_ice_candidate.borrow_mut().take();
+        self.on_ice_connection_state_changed.borrow_mut().take();
         self.peer.close();
     }
 }