+++ jason/src/rpc/websocket.rs
@@ -4,44 +4,66 @@
 
 use std::{cell::RefCell, convert::TryFrom, rc::Rc};
 
-use futures::{channel::oneshot, future};
-use macro_attr::*;
+use derive_more::{Display, From, Into};
+use futures::{
+    channel::{mpsc, oneshot},
+    future::{self, LocalBoxFuture},
+    stream::LocalBoxStream,
+};
 use medea_client_api_proto::{ClientMsg, ServerMsg};
-use newtype_derive::NewtypeFrom;
-use thiserror::*;
+use tracerr::Traced;
 use web_sys::{CloseEvent, Event, MessageEvent, WebSocket as SysWebSocket};
 
 use crate::{
-    rpc::CloseMsg,
-    utils::{EventListener, WasmErr},
+    rpc::{CloseMsg, RpcTransport},
+    utils::{EventListener, EventListenerBindError, JsCaused, JsError},
 };
 
 /// Errors that may occur when working with [`WebSocket`].
-#[derive(Debug, Error)]
-pub enum Error {
-    #[error("failed to create WebSocket: {0}")]
-    CreateSocket(WasmErr),
-    #[error("failed to init WebSocket")]
+#[derive(Debug, Display, JsCaused)]
+pub enum TransportError {
+    /// Occurs when the port to which the connection is being attempted
+    /// is being blocked.
+    #[display(fmt = "Failed to create WebSocket: {}", _0)]
+    CreateSocket(JsError),
+
+    /// Occurs when the connection close before becomes state active.
+    #[display(fmt = "Failed to init WebSocket")]
     InitSocket,
-    #[error("failed to parse client message: {0}")]
+
+    /// Occurs when [`ClientMessage`] cannot be parsed.
+    #[display(fmt = "Failed to parse client message: {}", _0)]
     ParseClientMessage(serde_json::error::Error),
-    #[error("failed to parse server message: {0}")]
+
+    /// Occurs when [`ServerMessage`] cannot be parsed.
+    #[display(fmt = "Failed to parse server message: {}", _0)]
     ParseServerMessage(serde_json::error::Error),
-    #[error("message is not a string")]
+
+    /// Occurs if the parsed message is not string.
+    #[display(fmt = "Message is not a string")]
     MessageNotString,
-    #[error("failed to send message: {0}")]
-    SendMessage(WasmErr),
-    #[error("failed to set handler for CloseEvent: {0}")]
-    SetHandlerOnClose(WasmErr),
-    #[error("failed to set handler for OpenEvent: {0}")]
-    SetHandlerOnOpen(WasmErr),
-    #[error("failed to set handler for MessageEvent: {0}")]
-    SetHandlerOnMessage(WasmErr),
-    #[error("could not cast {0} to State variant")]
-    CastState(u16),
-    #[error("underlying socket is closed")]
+
+    /// Occurs when a message cannot be send to server.
+    #[display(fmt = "Failed to send message: {}", _0)]
+    SendMessage(JsError),
+
+    /// Occurs when handler failed to bind to some [`WebSocket`] event. Not
+    /// really supposed to ever happen.
+    #[display(fmt = "Failed to bind to WebSocket event: {}", _0)]
+    WebSocketEventBindError(EventListenerBindError),
+
+    /// Occurs when message is sent to a closed socket.
+    #[display(fmt = "Underlying socket is closed")]
     ClosedSocket,
 }
+
+impl From<EventListenerBindError> for TransportError {
+    fn from(err: EventListenerBindError) -> Self {
+        Self::WebSocketEventBindError(err)
+    }
+}
+
+type Result<T, E = Traced<TransportError>> = std::result::Result<T, E>;
 
 /// State of websocket.
 #[derive(Debug)]
@@ -62,16 +84,14 @@ impl State {
     }
 }
 
-impl TryFrom<u16> for State {
-    type Error = Error;
-
-    fn try_from(value: u16) -> Result<Self, Self::Error> {
+impl From<u16> for State {
+    fn from(value: u16) -> Self {
         match value {
-            0 => Ok(Self::CONNECTING),
-            1 => Ok(Self::OPEN),
-            2 => Ok(Self::CLOSING),
-            3 => Ok(Self::CLOSED),
-            _ => Err(Error::CastState(value)),
+            0 => Self::CONNECTING,
+            1 => Self::OPEN,
+            2 => Self::CLOSING,
+            3 => Self::CLOSED,
+            _ => unreachable!(),
         }
     }
 }
@@ -85,13 +105,15 @@ struct InnerSocket {
     on_error: Option<EventListener<SysWebSocket, Event>>,
 }
 
-pub struct WebSocket(Rc<RefCell<InnerSocket>>);
+/// WebSocket [`RpcTransport`] between client and server.
+pub struct WebSocketRpcTransport(Rc<RefCell<InnerSocket>>);
 
 impl InnerSocket {
-    fn new(url: &str) -> Result<Self, Error> {
+    fn new(url: &str) -> Result<Self> {
         let socket = SysWebSocket::new(url)
             .map_err(Into::into)
-            .map_err(Error::CreateSocket)?;
+            .map_err(TransportError::CreateSocket)
+            .map_err(tracerr::wrap!())?;
         Ok(Self {
             socket_state: State::CONNECTING,
             socket: Rc::new(socket),
@@ -104,20 +126,85 @@ impl InnerSocket {
 
     /// Checks underlying WebSocket state and updates `socket_state`.
     fn update_state(&mut self) {
-        match State::try_from(self.socket.ready_state()) {
-            Ok(new_state) => self.socket_state = new_state,
-            Err(err) => {
-                // unreachable, unless some vendor will break enum
-                console_error!(err.to_string())
-            }
-        };
+        self.socket_state = self.socket.ready_state().into();
     }
 }
 
-impl WebSocket {
+impl RpcTransport for WebSocketRpcTransport {
+    fn on_message(&self) -> Result<LocalBoxStream<'static, Result<ServerMsg>>> {
+        let (tx, rx) = mpsc::unbounded();
+        let mut inner_mut = self.0.borrow_mut();
+        inner_mut.on_message = Some(
+            EventListener::new_mut(
+                Rc::clone(&inner_mut.socket),
+                "message",
+                move |msg| {
+                    let parsed = ServerMessage::try_from(&msg)
+                        .map(Into::into)
+                        .map_err(tracerr::wrap!());
+                    tx.unbounded_send(parsed).unwrap_or_else(|e| {
+                        console_error!(format!(
+                            "WebSocket's 'on_message' callback receiver \
+                             unexpectedly gone. {:?}",
+                            e
+                        ))
+                    });
+                },
+            )
+            .map_err(tracerr::map_from_and_wrap!(=> TransportError))?,
+        );
+        Ok(Box::pin(rx))
+    }
+
+    fn on_close(
+        &self,
+    ) -> Result<LocalBoxFuture<'static, Result<CloseMsg, oneshot::Canceled>>>
+    {
+        let (tx, rx) = oneshot::channel();
+        let mut inner_mut = self.0.borrow_mut();
+        let inner = Rc::clone(&self.0);
+        inner_mut.on_close = Some(
+            EventListener::new_once(
+                Rc::clone(&inner_mut.socket),
+                "close",
+                move |msg: CloseEvent| {
+                    inner.borrow_mut().update_state();
+                    tx.send(CloseMsg::from(&msg)).unwrap_or_else(|e| {
+                        console_error!(format!(
+                            "WebSocket's 'on_close' callback receiver \
+                             unexpectedly gone. {:?}",
+                            e
+                        ))
+                    });
+                },
+            )
+            .map_err(tracerr::map_from_and_wrap!(=> TransportError))?,
+        );
+        Ok(Box::pin(rx))
+    }
+
+    fn send(&self, msg: &ClientMsg) -> Result<()> {
+        let inner = self.0.borrow();
+        let message = serde_json::to_string(msg)
+            .map_err(TransportError::ParseClientMessage)
+            .map_err(tracerr::wrap!())?;
+
+        match inner.socket_state {
+            State::OPEN => inner
+                .socket
+                .send_with_str(&message)
+                .map_err(Into::into)
+                .map_err(TransportError::SendMessage)
+                .map_err(tracerr::wrap!()),
+            _ => Err(tracerr::new!(TransportError::ClosedSocket)),
+        }
+    }
+}
+
+impl WebSocketRpcTransport {
     /// Initiates new WebSocket connection. Resolves only when underlying
     /// connection becomes active.
-    pub async fn new(url: &str) -> Result<Self, Error> {
+    pub async fn new(url: &str) -> Result<Self> {
         let (tx_close, rx_close) = oneshot::channel();
         let (tx_open, rx_open) = oneshot::channel();
 
@@ -136,7 +223,7 @@ impl WebSocket {
                         let _ = tx_close.send(());
                     },
                 )
-                .map_err(Error::SetHandlerOnClose)?,
+                .map_err(tracerr::map_from_and_wrap!())?,
             );
 
             let inner = Rc::clone(&socket);
@@ -149,7 +236,7 @@ impl WebSocket {
                         let _ = tx_open.send(());
                     },
                 )
-                .map_err(Error::SetHandlerOnOpen)?,
+                .map_err(tracerr::map_from_and_wrap!(=> TransportError))?,
             );
         }
 
@@ -161,73 +248,16 @@ impl WebSocket {
         match state {
             future::Either::Left((opened, _)) => match opened {
                 Ok(_) => Ok(Self(socket)),
-                Err(_) => Err(Error::InitSocket),
+                Err(_) => Err(tracerr::new!(TransportError::InitSocket)),
             },
-            future::Either::Right(_closed) => Err(Error::InitSocket),
-        }
-    }
-
-    /// Set handler on receive message from server.
-    pub fn on_message<F>(&self, mut f: F) -> Result<(), Error>
-    where
-        F: (FnMut(Result<ServerMsg, Error>)) + 'static,
-    {
-        let mut inner_mut = self.0.borrow_mut();
-        inner_mut.on_message = Some(
-            EventListener::new_mut(
-                Rc::clone(&inner_mut.socket),
-                "message",
-                move |msg| {
-                    let parsed = ServerMessage::try_from(&msg)
-                        .map(std::convert::Into::into);
-                    f(parsed);
-                },
-            )
-            .map_err(Into::into)
-            .map_err(Error::SetHandlerOnMessage)?,
-        );
-        Ok(())
-    }
-
-    /// Set handler on close socket.
-    pub fn on_close<F>(&self, f: F) -> Result<(), Error>
-    where
-        F: (FnOnce(CloseMsg)) + 'static,
-    {
-        let mut inner_mut = self.0.borrow_mut();
-        let inner = Rc::clone(&self.0);
-        inner_mut.on_close = Some(
-            EventListener::new_once(
-                Rc::clone(&inner_mut.socket),
-                "close",
-                move |msg: CloseEvent| {
-                    inner.borrow_mut().update_state();
-                    f(CloseMsg::from(&msg));
-                },
-            )
-            .map_err(Error::SetHandlerOnClose)?,
-        );
-        Ok(())
-    }
-
-    /// Send message to server.
-    pub fn send(&self, msg: &ClientMsg) -> Result<(), Error> {
-        let inner = self.0.borrow();
-        let message =
-            serde_json::to_string(msg).map_err(Error::ParseClientMessage)?;
-
-        match inner.socket_state {
-            State::OPEN => inner
-                .socket
-                .send_with_str(&message)
-                .map_err(Into::into)
-                .map_err(Error::SendMessage),
-            _ => Err(Error::ClosedSocket),
+            future::Either::Right(_closed) => {
+                Err(tracerr::new!(TransportError::InitSocket))
+            }
         }
     }
 }
 
-impl Drop for WebSocket {
+impl Drop for WebSocketRpcTransport {
     fn drop(&mut self) {
         let mut inner = self.0.borrow_mut();
         if inner.socket_state.can_close() {
@@ -257,19 +287,21 @@ impl From<&CloseEvent> for CloseMsg {
     }
 }
 
-macro_attr! {
-    #[derive(NewtypeFrom!)]
-    pub struct ServerMessage(ServerMsg);
-}
+/// Wrapper for help to get [`ServerMsg`] from Websocket [MessageEvent][1].
+///
+/// [1]: https://developer.mozilla.org/en-US/docs/Web/API/MessageEvent
+#[derive(From, Into)]
+pub struct ServerMessage(ServerMsg);
 
 impl TryFrom<&MessageEvent> for ServerMessage {
-    type Error = Error;
+    type Error = TransportError;
 
     fn try_from(msg: &MessageEvent) -> std::result::Result<Self, Self::Error> {
-        let payload = msg.data().as_string().ok_or(Error::MessageNotString)?;
+        use TransportError::*;
+        let payload = msg.data().as_string().ok_or(MessageNotString)?;
 
         serde_json::from_str::<ServerMsg>(&payload)
-            .map_err(Error::ParseServerMessage)
+            .map_err(ParseServerMessage)
             .map(Self::from)
     }
 }