+++ jason/tests/media/manager.rs
@@ -1,16 +1,15 @@
 #![cfg(target_arch = "wasm32")]
 
-use futures::TryFutureExt;
 use js_sys::Array;
 use wasm_bindgen_futures::JsFuture;
 use wasm_bindgen_test::*;
 
 use medea_jason::{
-    media::MediaManager, utils::WasmErr, AudioTrackConstraints,
-    MediaStreamConstraints, VideoTrackConstraints,
+    media::MediaManager, AudioTrackConstraints, MediaStreamConstraints,
+    VideoTrackConstraints,
 };
 
-use crate::MockNavigator;
+use crate::{get_jason_error, MockNavigator};
 
 wasm_bindgen_test_configure!(run_in_browser);
 
@@ -29,26 +28,30 @@ async fn get_media_devices_info() {
 #[wasm_bindgen_test]
 async fn failed_get_media_devices_info() {
     let mock_navigator = MockNavigator::new();
-    mock_navigator.error_enumerate_devices("some error".into());
+    mock_navigator
+        .error_enumerate_devices("failed_get_media_devices_info".into());
     let media_manager = MediaManager::default();
-    let result = JsFuture::from(media_manager.new_handle().enumerate_devices())
-        .map_err(WasmErr::from)
-        .await;
+    let result =
+        JsFuture::from(media_manager.new_handle().enumerate_devices()).await;
     mock_navigator.stop();
     match result {
         Ok(_) => assert!(false),
-        Err(err) => assert_eq!(
-            err.to_string(),
-            "Error: MediaDevices.enumerateDevices() failed: some error"
-                .to_string()
-        ),
+        Err(err) => {
+            let e = get_jason_error(err);
+            assert_eq!(e.name(), "EnumerateDevicesFailed");
+            assert_eq!(
+                e.message(),
+                "MediaDevices.enumerateDevices() failed: Unknown JS error: \
+                 failed_get_media_devices_info",
+            );
+        }
     }
 }
 
 #[wasm_bindgen_test]
 async fn failed_get_user_media() {
     let mock_navigator = MockNavigator::new();
-    mock_navigator.error_get_user_media("some error".into());
+    mock_navigator.error_get_user_media("failed_get_user_media".into());
     let media_manager = MediaManager::default();
     let constraints = {
         let mut constraints = MediaStreamConstraints::new();
@@ -63,14 +66,56 @@ async fn failed_get_user_media() {
     let result = JsFuture::from(
         media_manager.new_handle().init_local_stream(constraints),
     )
-    .map_err(WasmErr::from)
     .await;
     mock_navigator.stop();
     match result {
         Ok(_) => assert!(false),
-        Err(err) => assert_eq!(
-            err.to_string(),
-            "Error: MediaDevices.getUserMedia() failed: some error".to_string()
-        ),
+        Err(err) => {
+            let err = get_jason_error(err);
+            assert_eq!(err.name(), "GetUserMediaFailed");
+            assert_eq!(
+                err.message(),
+                "MediaDevices.getUserMedia() failed: Unknown JS error: \
+                 failed_get_user_media",
+            );
+        }
+    }
+}
+
+#[wasm_bindgen_test]
+async fn failed_get_user_media2() {
+    let mock_navigator = MockNavigator::new();
+
+    let error = js_sys::Error::new("get_user_media_error_message");
+    error.set_name("get_user_media_error_name");
+
+    mock_navigator.error_get_user_media(error.into());
+    let media_manager = MediaManager::default();
+    let constraints = {
+        let mut constraints = MediaStreamConstraints::new();
+        let audio_constraints = AudioTrackConstraints::new();
+        let video_constraints = VideoTrackConstraints::new();
+
+        constraints.audio(audio_constraints);
+        constraints.video(video_constraints);
+
+        constraints
+    };
+    let result = JsFuture::from(
+        media_manager.new_handle().init_local_stream(constraints),
+    )
+    .await;
+    mock_navigator.stop();
+    match result {
+        Ok(_) => assert!(false),
+        Err(err) => {
+            let err = get_jason_error(err);
+            assert_eq!(err.name(), "GetUserMediaFailed");
+            assert_eq!(
+                err.message(),
+                "MediaDevices.getUserMedia() failed: \
+                 get_user_media_error_name: get_user_media_error_message",
+            );
+        }
     }
 }